Move aggregation job test code to `testing`
Add a new struct, `AggregationJobTest`, for testing the aggregation
flow. `AggregationJobTest::new()` generates task parameters for a given
VDAF config, DAP protocol version, and HPKE KEM algorithm for generating
receiver configs.

The code was moved out of `daphne/src/vdaf/mod_test.rs` where it was
called `Test`.
