+++ daphne/src/vdaf/mod_test.rs
@@ -5,17 +5,16 @@ use crate::{
     assert_metrics_include, assert_metrics_include_auxiliary_function, async_test_version,
     async_test_versions,
     error::DapAbort,
-    hpke::{HpkeAeadId, HpkeConfig, HpkeKdfId, HpkeKemId, HpkeReceiverConfig},
+    hpke::{HpkeAeadId, HpkeConfig, HpkeKdfId, HpkeKemId},
     messages::{
-        AggregationJobContinueReq, AggregationJobInitReq, AggregationJobResp, BatchSelector,
-        HpkeCiphertext, Interval, PartialBatchSelector, Report, ReportId, ReportShare, TaskId,
-        Time, Transition, TransitionFailure, TransitionVar,
+        AggregationJobInitReq, BatchSelector, Interval, PartialBatchSelector, Report, ReportId,
+        ReportShare, Transition, TransitionFailure, TransitionVar,
     },
-    metrics::DaphneMetrics,
-    test_version, test_versions, DapAggregateResult, DapAggregateShare, DapError, DapHelperState,
-    DapHelperTransition, DapLeaderState, DapLeaderTransition, DapLeaderUncommitted, DapMeasurement,
-    DapOutputShare, DapQueryConfig, DapTaskConfig, DapVersion, MetaAggregationJobId, Prio3Config,
-    VdafAggregateShare, VdafConfig, VdafMessage, VdafState,
+    test_version, test_versions,
+    testing::AggregationJobTest,
+    DapAggregateResult, DapAggregateShare, DapError, DapHelperState, DapHelperTransition,
+    DapLeaderState, DapLeaderTransition, DapLeaderUncommitted, DapMeasurement, DapOutputShare,
+    DapVersion, Prio3Config, VdafAggregateShare, VdafConfig, VdafMessage, VdafState,
 };
 use assert_matches::assert_matches;
 use hpke_rs::HpkePublicKey;
@@ -28,8 +27,7 @@ use prio::{
     },
 };
 use rand::prelude::*;
-use std::{fmt::Debug, time::SystemTime};
-use url::Url;
+use std::fmt::Debug;
 
 impl<M: Debug> DapLeaderTransition<M> {
     pub(crate) fn unwrap_continue(self) -> (DapLeaderState, M) {
@@ -75,7 +73,7 @@ impl<M: Debug> DapHelperTransition<M> {
 const TEST_VDAF: &VdafConfig = &VdafConfig::Prio3(Prio3Config::Count);
 
 async fn roundtrip_report(version: DapVersion) {
-    let t = Test::new(TEST_VDAF, version);
+    let t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let report = t
         .task_config
         .vdaf
@@ -153,7 +151,7 @@ async fn roundtrip_report(version: DapVersion) {
 async_test_versions! { roundtrip_report }
 
 fn roundtrip_report_unsupported_hpke_suite(version: DapVersion) {
-    let t = Test::new(TEST_VDAF, version);
+    let t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
 
     // The helper's HPKE config indicates a KEM type no supported by the client.
     let unsupported_hpke_config_list = vec![
@@ -183,7 +181,7 @@ fn roundtrip_report_unsupported_hpke_suite(version: DapVersion) {
 test_versions! { roundtrip_report_unsupported_hpke_suite }
 
 async fn produce_agg_job_init_req(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![
         DapMeasurement::U64(1),
         DapMeasurement::U64(0),
@@ -219,7 +217,7 @@ async fn produce_agg_job_init_req(version: DapVersion) {
 async_test_versions! { produce_agg_job_init_req }
 
 async fn produce_agg_job_init_req_skip_hpke_decrypt_err(version: DapVersion) {
-    let t = Test::new(TEST_VDAF, version);
+    let t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let mut reports = t.produce_reports(vec![DapMeasurement::U64(1)]);
 
     // Simulate HPKE decryption error of leader's report share.
@@ -238,7 +236,7 @@ async fn produce_agg_job_init_req_skip_hpke_decrypt_err(version: DapVersion) {
 async_test_versions! { produce_agg_job_init_req_skip_hpke_decrypt_err }
 
 async fn produce_agg_job_init_req_skip_hpke_unknown_config_id(version: DapVersion) {
-    let t = Test::new(TEST_VDAF, version);
+    let t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let mut reports = t.produce_reports(vec![DapMeasurement::U64(1)]);
 
     // Client tries to send Leader encrypted input with incorrect config ID.
@@ -257,7 +255,7 @@ async fn produce_agg_job_init_req_skip_hpke_unknown_config_id(version: DapVersio
 async_test_versions! { produce_agg_job_init_req_skip_hpke_unknown_config_id }
 
 async fn produce_agg_job_init_req_skip_vdaf_prep_error(version: DapVersion) {
-    let t = Test::new(TEST_VDAF, version);
+    let t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = vec![
         t.produce_invalid_report_public_share_decode_failure(DapMeasurement::U64(1), version),
         t.produce_invalid_report_input_share_decode_failure(DapMeasurement::U64(1), version),
@@ -276,7 +274,7 @@ async fn produce_agg_job_init_req_skip_vdaf_prep_error(version: DapVersion) {
 async_test_versions! { produce_agg_job_init_req_skip_vdaf_prep_error }
 
 async fn handle_agg_job_init_req_hpke_decrypt_err(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let mut reports = t.produce_reports(vec![DapMeasurement::U64(1)]);
 
     // Simulate HPKE decryption error of helper's report share.
@@ -302,7 +300,7 @@ async fn handle_agg_job_init_req_hpke_decrypt_err(version: DapVersion) {
 async_test_versions! { handle_agg_job_init_req_hpke_decrypt_err }
 
 async fn handle_agg_job_init_req_hpke_unknown_config_id(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let mut reports = t.produce_reports(vec![DapMeasurement::U64(1)]);
 
     // Client tries to send Helper encrypted input with incorrect config ID.
@@ -328,7 +326,7 @@ async fn handle_agg_job_init_req_hpke_unknown_config_id(version: DapVersion) {
 async_test_versions! { handle_agg_job_init_req_hpke_unknown_config_id }
 
 async fn handle_agg_job_init_req_vdaf_prep_error(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let report0 =
         t.produce_invalid_report_public_share_decode_failure(DapMeasurement::U64(1), version);
     let report1 =
@@ -373,7 +371,7 @@ async fn handle_agg_job_init_req_vdaf_prep_error(version: DapVersion) {
 async_test_versions! { handle_agg_job_init_req_vdaf_prep_error }
 
 async fn agg_job_resp_abort_transition_out_of_order(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![DapMeasurement::U64(1), DapMeasurement::U64(1)]);
     let (leader_state, agg_job_init_req) =
         t.produce_agg_job_init_req(reports).await.unwrap_continue();
@@ -396,7 +394,7 @@ async fn agg_job_resp_abort_transition_out_of_order(version: DapVersion) {
 async_test_versions! { agg_job_resp_abort_transition_out_of_order }
 
 async fn agg_job_resp_abort_report_id_repeated(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![DapMeasurement::U64(1), DapMeasurement::U64(1)]);
     let (leader_state, agg_job_init_req) =
         t.produce_agg_job_init_req(reports).await.unwrap_continue();
@@ -419,7 +417,7 @@ async_test_versions! { agg_job_resp_abort_report_id_repeated }
 
 async fn agg_job_resp_abort_unrecognized_report_id(version: DapVersion) {
     let mut rng = thread_rng();
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![DapMeasurement::U64(1), DapMeasurement::U64(1)]);
     let (leader_state, agg_job_init_req) =
         t.produce_agg_job_init_req(reports).await.unwrap_continue();
@@ -443,7 +441,7 @@ async fn agg_job_resp_abort_unrecognized_report_id(version: DapVersion) {
 async_test_versions! { agg_job_resp_abort_unrecognized_report_id }
 
 async fn agg_job_resp_abort_invalid_transition(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![DapMeasurement::U64(1)]);
     let (leader_state, agg_job_init_req) =
         t.produce_agg_job_init_req(reports).await.unwrap_continue();
@@ -464,7 +462,7 @@ async fn agg_job_resp_abort_invalid_transition(version: DapVersion) {
 async_test_versions! { agg_job_resp_abort_invalid_transition }
 
 async fn agg_job_cont_req(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![
         DapMeasurement::U64(1),
         DapMeasurement::U64(1),
@@ -528,7 +526,7 @@ async fn agg_job_cont_req(version: DapVersion) {
 async_test_versions! { agg_job_cont_req }
 
 async fn agg_job_cont_req_skip_vdaf_prep_error(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let mut reports = t.produce_reports(vec![DapMeasurement::U64(1), DapMeasurement::U64(1)]);
     reports.insert(
         1,
@@ -570,7 +568,7 @@ async_test_versions! { agg_job_cont_req_skip_vdaf_prep_error }
 
 async fn agg_cont_abort_unrecognized_report_id(version: DapVersion) {
     let mut rng = thread_rng();
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![DapMeasurement::U64(1), DapMeasurement::U64(1)]);
     let (leader_state, agg_job_init_req) =
         t.produce_agg_job_init_req(reports).await.unwrap_continue();
@@ -600,7 +598,7 @@ async fn agg_cont_abort_unrecognized_report_id(version: DapVersion) {
 async_test_versions! { agg_cont_abort_unrecognized_report_id }
 
 async fn agg_job_cont_req_abort_transition_out_of_order(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![DapMeasurement::U64(1), DapMeasurement::U64(1)]);
     let (leader_state, agg_job_init_req) =
         t.produce_agg_job_init_req(reports).await.unwrap_continue();
@@ -626,7 +624,7 @@ async fn agg_job_cont_req_abort_transition_out_of_order(version: DapVersion) {
 async_test_versions! { agg_job_cont_req_abort_transition_out_of_order }
 
 async fn agg_job_cont_req_abort_report_id_repeated(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![DapMeasurement::U64(1), DapMeasurement::U64(1)]);
     let (leader_state, agg_job_init_req) =
         t.produce_agg_job_init_req(reports).await.unwrap_continue();
@@ -651,7 +649,7 @@ async fn agg_job_cont_req_abort_report_id_repeated(version: DapVersion) {
 async_test_versions! { agg_job_cont_req_abort_report_id_repeated }
 
 async fn encrypted_agg_share(version: DapVersion) {
-    let t = Test::new(TEST_VDAF, version);
+    let t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let leader_agg_share = DapAggregateShare {
         report_count: 50,
         min_time: 1637359200,
@@ -695,7 +693,7 @@ async fn encrypted_agg_share(version: DapVersion) {
 async_test_versions! { encrypted_agg_share }
 
 async fn helper_state_serialization(version: DapVersion) {
-    let mut t = Test::new(TEST_VDAF, version);
+    let mut t = AggregationJobTest::new(TEST_VDAF, HpkeKemId::X25519HkdfSha256, version);
     let reports = t.produce_reports(vec![
         DapMeasurement::U64(1),
         DapMeasurement::U64(1),
@@ -718,92 +716,7 @@ async fn helper_state_serialization(version: DapVersion) {
 
 async_test_versions! { helper_state_serialization }
 
-pub(crate) struct Test {
-    now: Time,
-    task_id: TaskId,
-    agg_job_id: MetaAggregationJobId<'static>,
-    task_config: DapTaskConfig,
-    leader_hpke_receiver_config: HpkeReceiverConfig,
-    helper_hpke_receiver_config: HpkeReceiverConfig,
-    client_hpke_config_list: Vec<HpkeConfig>,
-    collector_hpke_receiver_config: HpkeReceiverConfig,
-    prometheus_registry: prometheus::Registry,
-    leader_metrics: DaphneMetrics,
-    helper_metrics: DaphneMetrics,
-}
-
-impl Test {
-    pub(crate) fn new(vdaf: &VdafConfig, version: DapVersion) -> Test {
-        let mut rng = thread_rng();
-        let now = SystemTime::now()
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .unwrap()
-            .as_secs();
-        let task_id = TaskId(rng.gen());
-        let agg_job_id = MetaAggregationJobId::gen_for_version(&version);
-        let vdaf_verify_key = vdaf.gen_verify_key();
-        let leader_hpke_receiver_config =
-            HpkeReceiverConfig::gen(rng.gen(), HpkeKemId::X25519HkdfSha256).unwrap();
-        let helper_hpke_receiver_config =
-            HpkeReceiverConfig::gen(rng.gen(), HpkeKemId::X25519HkdfSha256).unwrap();
-        let collector_hpke_receiver_config =
-            HpkeReceiverConfig::gen(rng.gen(), HpkeKemId::X25519HkdfSha256).unwrap();
-        let leader_hpke_config = leader_hpke_receiver_config.clone().config;
-        let helper_hpke_config = helper_hpke_receiver_config.clone().config;
-        let collector_hpke_config = collector_hpke_receiver_config.clone().config;
-        let prometheus_registry = prometheus::Registry::new();
-        let leader_metrics =
-            DaphneMetrics::register(&prometheus_registry, Some("test_leader")).unwrap();
-        let helper_metrics =
-            DaphneMetrics::register(&prometheus_registry, Some("test_helper")).unwrap();
-
-        Test {
-            now,
-            task_id,
-            agg_job_id,
-            leader_hpke_receiver_config,
-            helper_hpke_receiver_config,
-            client_hpke_config_list: vec![leader_hpke_config, helper_hpke_config],
-            collector_hpke_receiver_config,
-            task_config: DapTaskConfig {
-                version,
-                leader_url: Url::parse("http://leader.com").unwrap(),
-                helper_url: Url::parse("https://helper.org").unwrap(),
-                time_precision: 500,
-                expiration: now + 500,
-                min_batch_size: 10,
-                query: DapQueryConfig::TimeInterval,
-                vdaf: vdaf.clone(),
-                vdaf_verify_key,
-                collector_hpke_config,
-                taskprov: false,
-            },
-            prometheus_registry,
-            leader_metrics,
-            helper_metrics,
-        }
-    }
-
-    fn produce_reports(&self, measurements: Vec<DapMeasurement>) -> Vec<Report> {
-        let mut reports = Vec::with_capacity(measurements.len());
-
-        for measurement in measurements.into_iter() {
-            reports.push(
-                self.task_config
-                    .vdaf
-                    .produce_report(
-                        &self.client_hpke_config_list,
-                        self.now,
-                        &self.task_id,
-                        measurement,
-                        self.task_config.version,
-                    )
-                    .unwrap(),
-            );
-        }
-        reports
-    }
-
+impl AggregationJobTest {
     // Tweak the Helper's share so that decoding succeeds but preparation fails.
     fn produce_invalid_report_vdaf_prep_failure(
         &self,
@@ -887,257 +800,5 @@ impl Test {
                 version,
             )
             .unwrap()
-    }
-
-    async fn produce_agg_job_init_req(
-        &self,
-        reports: Vec<Report>,
-    ) -> DapLeaderTransition<AggregationJobInitReq> {
-        let metrics = self
-            .leader_metrics
-            .with_host(self.task_config.leader_url.host_str().unwrap());
-        self.task_config
-            .vdaf
-            .produce_agg_job_init_req(
-                &self.leader_hpke_receiver_config,
-                &self.task_id,
-                &self.task_config,
-                &self.agg_job_id,
-                &PartialBatchSelector::TimeInterval,
-                reports,
-                &metrics,
-            )
-            .await
-            .unwrap()
-    }
-
-    async fn handle_agg_job_init_req(
-        &mut self,
-        agg_job_init_req: AggregationJobInitReq,
-    ) -> DapHelperTransition<AggregationJobResp> {
-        let metrics = self
-            .helper_metrics
-            .with_host(self.task_config.helper_url.host_str().unwrap());
-        self.task_config
-            .vdaf
-            .handle_agg_job_init_req(
-                &self.helper_hpke_receiver_config,
-                &self.task_id,
-                &self.task_config,
-                &agg_job_init_req,
-                &metrics,
-            )
-            .await
-            .unwrap()
-    }
-
-    fn handle_agg_job_resp(
-        &self,
-        leader_state: DapLeaderState,
-        agg_job_resp: AggregationJobResp,
-    ) -> DapLeaderTransition<AggregationJobContinueReq> {
-        let metrics = self
-            .leader_metrics
-            .with_host(self.task_config.leader_url.host_str().unwrap());
-        self.task_config
-            .vdaf
-            .handle_agg_job_resp(
-                &self.task_id,
-                &self.agg_job_id,
-                leader_state,
-                agg_job_resp,
-                self.task_config.version,
-                &metrics,
-            )
-            .unwrap()
-    }
-
-    fn handle_agg_job_resp_expect_err(
-        &self,
-        leader_state: DapLeaderState,
-        agg_job_resp: AggregationJobResp,
-    ) -> DapAbort {
-        let metrics = self
-            .leader_metrics
-            .with_host(self.task_config.leader_url.host_str().unwrap());
-        self.task_config
-            .vdaf
-            .handle_agg_job_resp(
-                &self.task_id,
-                &self.agg_job_id,
-                leader_state,
-                agg_job_resp,
-                self.task_config.version,
-                &metrics,
-            )
-            .expect_err("handle_agg_job_resp() succeeded; expected failure")
-    }
-
-    fn handle_agg_job_cont_req(
-        &self,
-        helper_state: DapHelperState,
-        agg_job_cont_req: &AggregationJobContinueReq,
-    ) -> DapHelperTransition<AggregationJobResp> {
-        let metrics = self
-            .helper_metrics
-            .with_host(self.task_config.helper_url.host_str().unwrap());
-        self.task_config
-            .vdaf
-            .handle_agg_job_cont_req(
-                &self.task_id,
-                &self.agg_job_id,
-                helper_state,
-                agg_job_cont_req,
-                &metrics,
-            )
-            .unwrap()
-    }
-
-    fn handle_agg_job_cont_req_expect_err(
-        &self,
-        helper_state: DapHelperState,
-        agg_job_cont_req: &AggregationJobContinueReq,
-    ) -> DapAbort {
-        let metrics = self
-            .helper_metrics
-            .with_host(self.task_config.helper_url.host_str().unwrap());
-        self.task_config
-            .vdaf
-            .handle_agg_job_cont_req(
-                &self.task_id,
-                &self.agg_job_id,
-                helper_state,
-                agg_job_cont_req,
-                &metrics,
-            )
-            .expect_err("handle_agg_job_cont_req() succeeded; expected failure")
-    }
-
-    fn handle_final_agg_job_resp(
-        &self,
-        leader_uncommitted: DapLeaderUncommitted,
-        agg_job_resp: AggregationJobResp,
-    ) -> Vec<DapOutputShare> {
-        let metrics = self
-            .leader_metrics
-            .with_host(self.task_config.leader_url.host_str().unwrap());
-        self.task_config
-            .vdaf
-            .handle_final_agg_job_resp(leader_uncommitted, agg_job_resp, &metrics)
-            .unwrap()
-    }
-
-    fn produce_leader_encrypted_agg_share(
-        &self,
-        batch_selector: &BatchSelector,
-        agg_share: &DapAggregateShare,
-    ) -> HpkeCiphertext {
-        self.task_config
-            .vdaf
-            .produce_leader_encrypted_agg_share(
-                &self.task_config.collector_hpke_config,
-                &self.task_id,
-                batch_selector,
-                agg_share,
-                self.task_config.version,
-            )
-            .unwrap()
-    }
-
-    fn produce_helper_encrypted_agg_share(
-        &self,
-        batch_selector: &BatchSelector,
-        agg_share: &DapAggregateShare,
-    ) -> HpkeCiphertext {
-        self.task_config
-            .vdaf
-            .produce_helper_encrypted_agg_share(
-                &self.task_config.collector_hpke_config,
-                &self.task_id,
-                batch_selector,
-                agg_share,
-                self.task_config.version,
-            )
-            .unwrap()
-    }
-
-    async fn consume_encrypted_agg_shares(
-        &self,
-        batch_selector: &BatchSelector,
-        report_count: u64,
-        enc_agg_shares: Vec<HpkeCiphertext>,
-    ) -> DapAggregateResult {
-        self.task_config
-            .vdaf
-            .consume_encrypted_agg_shares(
-                &self.collector_hpke_receiver_config,
-                &self.task_id,
-                batch_selector,
-                report_count,
-                enc_agg_shares,
-                self.task_config.version,
-            )
-            .await
-            .unwrap()
-    }
-
-    pub(crate) async fn roundtrip(
-        &mut self,
-        measurements: Vec<DapMeasurement>,
-    ) -> DapAggregateResult {
-        let batch_selector = BatchSelector::TimeInterval {
-            batch_interval: Interval {
-                start: self.now,
-                duration: 3600,
-            },
-        };
-
-        // Clients: Shard
-        let reports = self.produce_reports(measurements);
-
-        // Aggregators: Preparation
-        let (leader_state, agg_init) = self
-            .produce_agg_job_init_req(reports)
-            .await
-            .unwrap_continue();
-        let (helper_state, agg_job_resp) = self
-            .handle_agg_job_init_req(agg_init)
-            .await
-            .unwrap_continue();
-        let got = DapHelperState::get_decoded(
-            &self.task_config.vdaf,
-            &helper_state
-                .get_encoded(&self.task_config.vdaf)
-                .expect("failed to encode helper state"),
-        )
-        .expect("failed to decode helper state");
-        assert_eq!(got, helper_state);
-
-        let (uncommitted, agg_cont) = self
-            .handle_agg_job_resp(leader_state, agg_job_resp)
-            .unwrap_uncommitted();
-        let (helper_out_shares, agg_job_resp) = self
-            .handle_agg_job_cont_req(helper_state, &agg_cont)
-            .unwrap_finish();
-        let leader_out_shares = self.handle_final_agg_job_resp(uncommitted, agg_job_resp);
-        let report_count = u64::try_from(leader_out_shares.len()).unwrap();
-
-        // Leader: Aggregation
-        let leader_agg_share = DapAggregateShare::try_from_out_shares(leader_out_shares).unwrap();
-        let leader_encrypted_agg_share =
-            self.produce_leader_encrypted_agg_share(&batch_selector, &leader_agg_share);
-
-        // Helper: Aggregation
-        let helper_agg_share = DapAggregateShare::try_from_out_shares(helper_out_shares).unwrap();
-        let helper_encrypted_agg_share =
-            self.produce_helper_encrypted_agg_share(&batch_selector, &helper_agg_share);
-
-        // Collector: Unshard
-        self.consume_encrypted_agg_shares(
-            &batch_selector,
-            report_count,
-            vec![leader_encrypted_agg_share, helper_encrypted_agg_share],
-        )
-        .await
     }
 }
